+++ specado-node/index.js
@@ -78,8 +78,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-android-arm64')
-        const bindingPackageVersion = require('@specado/core-android-arm64/package.json').version
+        const binding = require('specado-android-arm64')
+        const bindingPackageVersion = require('specado-android-arm64/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -94,8 +94,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-android-arm-eabi')
-        const bindingPackageVersion = require('@specado/core-android-arm-eabi/package.json').version
+        const binding = require('specado-android-arm-eabi')
+        const bindingPackageVersion = require('specado-android-arm-eabi/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -114,8 +114,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-win32-x64-msvc')
-        const bindingPackageVersion = require('@specado/core-win32-x64-msvc/package.json').version
+        const binding = require('specado-win32-x64-msvc')
+        const bindingPackageVersion = require('specado-win32-x64-msvc/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -130,8 +130,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-win32-ia32-msvc')
-        const bindingPackageVersion = require('@specado/core-win32-ia32-msvc/package.json').version
+        const binding = require('specado-win32-ia32-msvc')
+        const bindingPackageVersion = require('specado-win32-ia32-msvc/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -146,8 +146,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-win32-arm64-msvc')
-        const bindingPackageVersion = require('@specado/core-win32-arm64-msvc/package.json').version
+        const binding = require('specado-win32-arm64-msvc')
+        const bindingPackageVersion = require('specado-win32-arm64-msvc/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -165,8 +165,8 @@ function requireNative() {
       loadErrors.push(e)
     }
     try {
-      const binding = require('@specado/core-darwin-universal')
-      const bindingPackageVersion = require('@specado/core-darwin-universal/package.json').version
+      const binding = require('specado-darwin-universal')
+      const bindingPackageVersion = require('specado-darwin-universal/package.json').version
       if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
         throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
       }
@@ -181,8 +181,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-darwin-x64')
-        const bindingPackageVersion = require('@specado/core-darwin-x64/package.json').version
+        const binding = require('specado-darwin-x64')
+        const bindingPackageVersion = require('specado-darwin-x64/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -197,8 +197,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-darwin-arm64')
-        const bindingPackageVersion = require('@specado/core-darwin-arm64/package.json').version
+        const binding = require('specado-darwin-arm64')
+        const bindingPackageVersion = require('specado-darwin-arm64/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -217,8 +217,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-freebsd-x64')
-        const bindingPackageVersion = require('@specado/core-freebsd-x64/package.json').version
+        const binding = require('specado-freebsd-x64')
+        const bindingPackageVersion = require('specado-freebsd-x64/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -233,8 +233,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-freebsd-arm64')
-        const bindingPackageVersion = require('@specado/core-freebsd-arm64/package.json').version
+        const binding = require('specado-freebsd-arm64')
+        const bindingPackageVersion = require('specado-freebsd-arm64/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -254,8 +254,8 @@ function requireNative() {
           loadErrors.push(e)
         }
         try {
-          const binding = require('@specado/core-linux-x64-musl')
-          const bindingPackageVersion = require('@specado/core-linux-x64-musl/package.json').version
+          const binding = require('specado-linux-x64-musl')
+          const bindingPackageVersion = require('specado-linux-x64-musl/package.json').version
           if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
             throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
           }
@@ -270,8 +270,8 @@ function requireNative() {
           loadErrors.push(e)
         }
         try {
-          const binding = require('@specado/core-linux-x64-gnu')
-          const bindingPackageVersion = require('@specado/core-linux-x64-gnu/package.json').version
+          const binding = require('specado-linux-x64-gnu')
+          const bindingPackageVersion = require('specado-linux-x64-gnu/package.json').version
           if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
             throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
           }
@@ -288,8 +288,8 @@ function requireNative() {
           loadErrors.push(e)
         }
         try {
-          const binding = require('@specado/core-linux-arm64-musl')
-          const bindingPackageVersion = require('@specado/core-linux-arm64-musl/package.json').version
+          const binding = require('specado-linux-arm64-musl')
+          const bindingPackageVersion = require('specado-linux-arm64-musl/package.json').version
           if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
             throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
           }
@@ -304,8 +304,8 @@ function requireNative() {
           loadErrors.push(e)
         }
         try {
-          const binding = require('@specado/core-linux-arm64-gnu')
-          const bindingPackageVersion = require('@specado/core-linux-arm64-gnu/package.json').version
+          const binding = require('specado-linux-arm64-gnu')
+          const bindingPackageVersion = require('specado-linux-arm64-gnu/package.json').version
           if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
             throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
           }
@@ -322,8 +322,8 @@ function requireNative() {
           loadErrors.push(e)
         }
         try {
-          const binding = require('@specado/core-linux-arm-musleabihf')
-          const bindingPackageVersion = require('@specado/core-linux-arm-musleabihf/package.json').version
+          const binding = require('specado-linux-arm-musleabihf')
+          const bindingPackageVersion = require('specado-linux-arm-musleabihf/package.json').version
           if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
             throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
           }
@@ -338,8 +338,8 @@ function requireNative() {
           loadErrors.push(e)
         }
         try {
-          const binding = require('@specado/core-linux-arm-gnueabihf')
-          const bindingPackageVersion = require('@specado/core-linux-arm-gnueabihf/package.json').version
+          const binding = require('specado-linux-arm-gnueabihf')
+          const bindingPackageVersion = require('specado-linux-arm-gnueabihf/package.json').version
           if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
             throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
           }
@@ -356,8 +356,8 @@ function requireNative() {
           loadErrors.push(e)
         }
         try {
-          const binding = require('@specado/core-linux-riscv64-musl')
-          const bindingPackageVersion = require('@specado/core-linux-riscv64-musl/package.json').version
+          const binding = require('specado-linux-riscv64-musl')
+          const bindingPackageVersion = require('specado-linux-riscv64-musl/package.json').version
           if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
             throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
           }
@@ -372,8 +372,8 @@ function requireNative() {
           loadErrors.push(e)
         }
         try {
-          const binding = require('@specado/core-linux-riscv64-gnu')
-          const bindingPackageVersion = require('@specado/core-linux-riscv64-gnu/package.json').version
+          const binding = require('specado-linux-riscv64-gnu')
+          const bindingPackageVersion = require('specado-linux-riscv64-gnu/package.json').version
           if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
             throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
           }
@@ -389,8 +389,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-linux-ppc64-gnu')
-        const bindingPackageVersion = require('@specado/core-linux-ppc64-gnu/package.json').version
+        const binding = require('specado-linux-ppc64-gnu')
+        const bindingPackageVersion = require('specado-linux-ppc64-gnu/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -405,8 +405,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-linux-s390x-gnu')
-        const bindingPackageVersion = require('@specado/core-linux-s390x-gnu/package.json').version
+        const binding = require('specado-linux-s390x-gnu')
+        const bindingPackageVersion = require('specado-linux-s390x-gnu/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -425,8 +425,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-openharmony-arm64')
-        const bindingPackageVersion = require('@specado/core-openharmony-arm64/package.json').version
+        const binding = require('specado-openharmony-arm64')
+        const bindingPackageVersion = require('specado-openharmony-arm64/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -441,8 +441,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-openharmony-x64')
-        const bindingPackageVersion = require('@specado/core-openharmony-x64/package.json').version
+        const binding = require('specado-openharmony-x64')
+        const bindingPackageVersion = require('specado-openharmony-x64/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -457,8 +457,8 @@ function requireNative() {
         loadErrors.push(e)
       }
       try {
-        const binding = require('@specado/core-openharmony-arm')
-        const bindingPackageVersion = require('@specado/core-openharmony-arm/package.json').version
+        const binding = require('specado-openharmony-arm')
+        const bindingPackageVersion = require('specado-openharmony-arm/package.json').version
         if (bindingPackageVersion !== '0.1.0' && process.env.NAPI_RS_ENFORCE_VERSION_CHECK && process.env.NAPI_RS_ENFORCE_VERSION_CHECK !== '0') {
           throw new Error(`Native binding package version mismatch, expected 0.1.0 but got ${bindingPackageVersion}. You can reinstall dependencies to fix this issue.`)
         }
@@ -486,7 +486,7 @@ if (!nativeBinding || process.env.NAPI_RS_FORCE_WASI) {
   }
   if (!nativeBinding) {
     try {
-      nativeBinding = require('@specado/core-wasm32-wasi')
+      nativeBinding = require('specado-wasm32-wasi')
     } catch (err) {
       if (process.env.NAPI_RS_FORCE_WASI) {
         loadErrors.push(err)