+++ specado-node/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-All notable changes to the @specado/core Node.js package will be documented in this file.
+All notable changes to the specado Node.js package will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -85,7 +85,7 @@ This release represents the Week 2 MVP of the Specado project, focusing on:
 ## Links
 - [Repository](https://github.com/specado/specado)
 - [Issues](https://github.com/specado/specado/issues)
-- [npm Package](https://www.npmjs.com/package/@specado/core)
+- [npm Package](https://www.npmjs.com/package/specado)
 - [Documentation](https://docs.specado.com)
 
 [Unreleased]: https://github.com/specado/specado/compare/v0.1.0...HEAD